+++ client/modules/news/actions/customActions.js
@@ -1,0 +1,16 @@
+/*****
+
+CUSTOM ACTIONS GO HERE
+
+i.e. non-CRUD actions like toggling tabs or filters 
+
+*****/
+
+export const SET_VISIBILITY_FILTER = "SET_VISIBILITY_FILTER";
+
+export const setVisibilityFilter = (filter) => {
+  return {
+    type: SET_VISIBILITY_FILTER
+    , filter
+  }
+}
