+++ client/modules/news/containers/Layout.js.jsx
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import Base from "../../../global/components/BaseComponent.js.jsx";
 import { bindActionCreators } from 'redux'
-import * as NewsActions from '../actions';
+
 
 import { connect } from 'react-redux';
 // import { push }
