+++ server/initialState.js
@@ -1,0 +1,10 @@
+exports.initialState = {
+
+    news: {
+      isFetching: true
+      , items: []
+    }
+  }
+
+
+// export default initialStateObj;
